# api/gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Defines values for ClusterStatus.
const (
	HEALTHERR  ClusterStatus = "HEALTH_ERR"
	HEALTHOK   ClusterStatus = "HEALTH_OK"
	HEALTHWARN ClusterStatus = "HEALTH_WARN"
)

// Cluster defines model for Cluster.
type Cluster struct {
	// Detail cluster 상태가 좋지 않을때, 해당 이유에 대해 보여준다.
	Detail *string `json:"detail,omitempty"`
	Id     string  `json:"id"`

	// IsStable 일정 시간 이상 HEALTH_OK가 유지되는 상태
	IsStable bool          `json:"is_stable"`
	Name     string        `json:"name"`
	Status   ClusterStatus `json:"status"`
}

// ClusterStatus defines model for Cluster.Status.
type ClusterStatus string

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// RegisterCluster defines model for RegisterCluster.
type RegisterCluster struct {
	Hosts []string `json:"hosts"`
	Key   string   `json:"key"`
	Name  string   `json:"name"`
}

// RegisterClusterJSONRequestBody defines body for RegisterCluster for application/json ContentType.
type RegisterClusterJSONRequestBody = RegisterCluster

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /clusters)
	ListClusters(w http.ResponseWriter, r *http.Request)

	// (POST /clusters)
	RegisterCluster(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (GET /clusters)
func (_ Unimplemented) ListClusters(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /clusters)
func (_ Unimplemented) RegisterCluster(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListClusters operation middleware
func (siw *ServerInterfaceWrapper) ListClusters(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListClusters(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RegisterCluster operation middleware
func (siw *ServerInterfaceWrapper) RegisterCluster(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RegisterCluster(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/clusters", wrapper.ListClusters)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/clusters", wrapper.RegisterCluster)
	})

	return r
}

type ListClustersRequestObject struct {
}

type ListClustersResponseObject interface {
	VisitListClustersResponse(w http.ResponseWriter) error
}

type ListClusters200JSONResponse []Cluster

func (response ListClusters200JSONResponse) VisitListClustersResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListClusters204Response struct {
}

func (response ListClusters204Response) VisitListClustersResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type ListClusters500JSONResponse Error

func (response ListClusters500JSONResponse) VisitListClustersResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type RegisterClusterRequestObject struct {
	Body *RegisterClusterJSONRequestBody
}

type RegisterClusterResponseObject interface {
	VisitRegisterClusterResponse(w http.ResponseWriter) error
}

type RegisterCluster201JSONResponse Cluster

func (response RegisterCluster201JSONResponse) VisitRegisterClusterResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type RegisterCluster400Response struct {
}

func (response RegisterCluster400Response) VisitRegisterClusterResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type RegisterCluster500JSONResponse Error

func (response RegisterCluster500JSONResponse) VisitRegisterClusterResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (GET /clusters)
	ListClusters(ctx context.Context, request ListClustersRequestObject) (ListClustersResponseObject, error)

	// (POST /clusters)
	RegisterCluster(ctx context.Context, request RegisterClusterRequestObject) (RegisterClusterResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// ListClusters operation middleware
func (sh *strictHandler) ListClusters(w http.ResponseWriter, r *http.Request) {
	var request ListClustersRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListClusters(ctx, request.(ListClustersRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListClusters")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListClustersResponseObject); ok {
		if err := validResponse.VisitListClustersResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// RegisterCluster operation middleware
func (sh *strictHandler) RegisterCluster(w http.ResponseWriter, r *http.Request) {
	var request RegisterClusterRequestObject

	var body RegisterClusterJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.RegisterCluster(ctx, request.(RegisterClusterRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RegisterCluster")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(RegisterClusterResponseObject); ok {
		if err := validResponse.VisitRegisterClusterResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}